Modified SimpleSAMLphp library
Had to change internal Utilities and Message classes because of
the way Laravel generates URL's. Since Laravel's urls all technically
throw the user back to index.php and route from there the url's
that SimpleSAMLphp was seeing were incorrect. As a result the urls
are now required as a parameter for this library. This is an
important change until we can come up with a more generic solution.